Correcting bug report #1 on Makefile
To correct this bug, I changed the conditions to check if the folder exist and if its not empty, the old version check if version.h  and the autoconf.h exist in the $$kernel_source folder. Another change I made is put a new check for the folder /usr/src/kernels, because on CentOS 7 this is the folder where it keeps the kernel sources.